Add new API.
This commit adds the first version of the new API, which removes some
unnecessary members and adds support for sending players to other
players, as well as some general information methods.

The new API is separated from the main ServerClusters class to reduce
clutter and make the documentation easier to edit and read.