Remove unnecessary shutdown function.
This commit removes the unnecessary shutdown function. This means that
servers will remove themselves from consideration in two ways: firstly,
if they do not want players to be sent to them, they can just reduce the
total number of slots through the API, which has the exact same
behavior, and reducing the duplication makes maintenance easier and bugs
less likely. Secondly, to be removed from other servers' caches
altogether, the servers just need to stop sending heartbeat messages and
they will be taken out. Similarly, this has almost the same behavior
without the need for a special messaging channel or protocol. 

The only real sacrifice in removing the shutdown message is that during
a "soft shutdown", where a server stops accepting new players for X
minutes before shutting down. Without a definitive shutdown message, the
server will still be sending out heartbeat messages until it shuts down.
However, this is not a major sacrifice and is not currently worth the
overhead and clutter in the API.