Remove "opt-out" option.
This is a major commit which removes the option for Minecraft servers
using ServerClusters to opt out of being a "p2p-server" on the
ServerClusters network. Furthermore, this commit seeks to take out
almost all references to "p2p-server", which is confusing and
unnecessary terminology.

This commit shifts ServerClusters to forcing servers to use it, to send
out heartbeats, and to declare themselves as part of a cluster. However,
a cluster is really just a logical identity of server resources. It is
perfectly acceptable for a server to be the only member of a cluster,
and basically not be a cluster at all. In such a case, all that
ServerClusters is forcing on them is that they report their status to
the network with heartbeat messages, and that ServerClusters manages
their player slots, which may or may not have any visible effect
whatsoerver, depending on what type of server it is.

This commit also makes a lot of edits to javadocs to remove ambiguous
documentation and terminology, although more is probably necessary in
the future. 

This commit does not yet edit the configuration of ServerClusters, but
just the usage of that configuration, ignoring the setting of whether
the server should be a "p2p-server" or not. (if this is confusing to
you, it just sort of shows that it was justified to get rid of this
distinction)