Reimplement reservation protocol.
This is a major commit that reimplements the way reservation requests
and responses work. Importantly, preliminary support for targeting
requests by player is added, although the API for these features has not
been fully implemented yet.

Additionally, the requests and responses are no longer manually
serialized/deserialized with byte buffer and now use Java-Object
serialization through apache SerializationUtils (heartbeat messages are
left untouched because they are very simple and need to be manually
edited after being created). Manual serialization was largely to help me
experiment with and understand serialization better, but it is
insufficient for more complex and changeable protocols.