Implement sending to players.
First implementation of sending players to another player on the
network. This commit adds the local behavior that takes advantage of the
new protocol, such as sending a request, how long to wait for a
response, and what to do when a response is received.

This feature is currently untested and likely will need to be tweaked
after testing.