Use ServerStatus instead of String.
This commit changes the return type of the method getServers() in
NetworkStatus. Rather than returning just the ids of the servers, this
commit makes it return the server status objects themselves. This commit
also updates the clients that use getServers() to use this new format.

By returning the ServerStatus objects, this opens up clients of
NetworkStatus to do a reasonable amount of their own analysis of the
network, allowing for more possible feature without too much hackiness.