Add sending players to servers.
This commit adds the ability to PlayerRelocator to try to send players
to a predefined set of servers, in addition to being able to send them
to a cluster or another player. 

This feature is not intended ever to be available through the API, but
enables a number of internally more sophisticated features for dealing
with common-sense issues like matchmaking.

Also fixes a forgotten method rename in PlayerRelocator.