Readd ShutdownNotification.
This commit readds shutdown notifications that will be sent when servers
shutdown in a clean way. This allows a listener on the network to
distinguish between a server shutting down cleanly and crashing with
much greater accuracy.

This commit also adds a new API method, closeThisServer() which gives a
client a more definitive way to make sure that no more players are sent
to this server until it shuts down.

Also removes a very spammy and no longer necessary heartbeat debug
message.