Update MC, exclude version from jar.
This commit changes build settings so that the final jar's name will not
contain the build's version. While in theory including versions are a
good thing, most of our changes are implemented through patches, where
overwriting an old jar with a new one is the primary way to update jars.
As such, having jars with different names just discourages incrementing
the version number at all, because it makes patches harder and more
prone to error.