Add new nodes, features, and improvements.
This massive commit adds a multitude of new features, fundamentally
restructures the project, and makes a number of improvements. This is a
major revision of ServerClusters that changes the networks structure and
behavior. This update will definitely break compatibility with just
about every client, but will mostly just require recompiling rather than
any logical changes.

Most aspects of the ServerClusters network were changed in some way,
with the biggest exception being the tested and proven
instance-provision system. Important changes include:
- Added new network nodes/clients for use on bungee instances and a
centralized (but low-traffic) network coordinator. 
- Restructured the project into a multi-module Maven project that builds
each node based on the same common protocol and core functions. Each
module represents an implementation of a node on a given platform
(bukkit and bungeecord respectively, for now).
- Bungeecord instances can now dynamically add and remove Minecraft
servers as they announce themselves and shutdown.
- Unique and human-readable server ids are now dynamically allocated by
the central network coordinator. 
- Many code improvements, including making the main ServerClusters
game-server client more generic, able to support multiple server
implementations simultaneously (though only a Bukkit client has been
implemented so far)
- Improved, added, and edited tons of documentation.
- Made network cache track the newly available information about servers
like their ip and port.
- Much of the configuration has been centralized to the network
coordinator, which will periodically serve requests and push updates to
the nodes on the network. While the config can dynamically update across
the network, currently not every feature will use the config in a
dynamically updating way. For example, if an admin were to change what
messaging channels are used for different functions, and the coordinator
pushed an update to all of the nodes, the nodes would not be able to
change the messaging channels they use until they restarted. This may
change in the future, but for now that is how it works and it is hard to
see a situation in which some configuration features would be dynamic,
given how unimportant dynamism is for them and how difficult it would be
to implement.
- The central network coordinator listens for servers that stop sending
heartbeats without first sending a shutdown message, and can send emails
to administrators to notify them of the possible crash.
- Improved appearance and efficiency of the networkstatus command.
- Replace apache SerializationUtils with a small internal serialization
library, because apache's refused to work on spigot servers in certain
circumstances.
- Allow gameserver singleton to be accessed while enabling (instead of
after).
- Split PlayerRelocator into separate server and client modules. 
- Implement server-clusters selection behavior in bungeecord when a
player first joins the network.
- Made a few long-overdue updates to the Readme (still more is needed).
- Updated .gitignore to be more comprehensive.
- Updated PubSub version.